+++ lib/gg_router.dart
@@ -12,3 +12,5 @@ export 'src/gg_route_information_parser.dart';
 export 'src/gg_router.dart';
 export 'src/gg_popover_route.dart';
 export 'src/gg_route_params.dart';
+
+export 'src/gg_route_change_builder.dart';