+++ test/gg_route_tree_node_test.dart
@@ -1326,6 +1326,11 @@ main() {
         root.findOrCreateParam(name: 'unknownParam', seed: 5);
         expect(root.param('unknownParam')?.value, 123);
       });
+
+      test('should parse semantic label', () {
+        root.json = '{"__semanticLabel": "Semantic Label"}';
+        expect(root.semanticLabel, 'Semantic Label');
+      });
     });
 
     // #########################################################################
@@ -1374,34 +1379,63 @@ main() {
         rootCopy.json = json;
         expect(rootCopy.stagedChild, rootCopy.findOrCreateChild('stagedChild'));
       });
-    });
 
-    // #########################################################################
-    group('set and get semanticsLabel', () {
-      test('should return the node name if no semantics label has been set',
-          () {
-        init();
-        expect(childC.semanticsLabel, childC.name);
-        dispose();
-      });
-
-      test('should return the previously set semantics label', () {
-        init();
-        childC.semanticsLabel = 'Child C';
-        expect(childC.semanticsLabel, 'Child C');
-        dispose();
+      test('should also save semantic label', () {
+        final root = GgRouteTreeNode(name: '_ROOT_');
+        root.semanticLabel = 'ABC';
+        final json = root.json;
+        final rootCopy = GgRouteTreeNode(name: '_ROOT_');
+        rootCopy.json = json;
+        expect(rootCopy.semanticLabel, 'ABC');
       });
     });
 
     // #########################################################################
-    group('semanticsLabelForPath', () {
-      test('should return the semantics label of the element relative to path',
+    group('set and get semanticLabel', () {
+      test('should return the node name if no semantic label has been set', () {
+        init();
+        expect(childC.semanticLabel, childC.name);
+        dispose();
+      });
+
+      test('should return the previously set semantic label', () {
+        init();
+        childC.semanticLabel = 'Child C';
+        expect(childC.semanticLabel, 'Child C');
+        dispose();
+      });
+
+      test(
+          'should return the the parent label for _INDEX_ routes '
+          'but only when index route has not semantic label set', () {
+        init();
+        final index = childA0.findOrCreateChild('_INDEX_');
+        expect(index.semanticLabel, childA0.semanticLabel);
+        index.semanticLabel = 'Semantic Label';
+        expect(index.semanticLabel, 'Semantic Label');
+        dispose();
+      });
+    });
+
+    // #########################################################################
+    group('semanticLabelForPath', () {
+      test('should return the semantic label of the element relative to path',
           () {
         init();
-        expect(childC.semanticsLabelForPath('.'), childC.name);
-        expect(childC.semanticsLabelForPath('..'), childB.name);
-        expect(childC.semanticsLabelForPath('/'), root.name);
-        expect(childC.semanticsLabelForPath('../../child-a0'), childA0.name);
+        expect(childC.semanticLabelForPath('.'), childC.name);
+        expect(childC.semanticLabelForPath('..'), childB.name);
+        expect(childC.semanticLabelForPath('/'), root.name);
+        expect(childC.semanticLabelForPath('../../'), childA0.name);
+        dispose();
+      });
+    });
+
+    // #########################################################################
+    group('setSemanticLabelForPath(path, label)', () {
+      test('should set the semantic label of the element relative to path', () {
+        init();
+        childC.setSemanticLabelForPath(path: '../../', label: 'DEF');
+        expect(childA0.semanticLabel, 'DEF');
         dispose();
       });
     });